+++ salt/output/key.py
@@ -8,9 +8,9 @@ The ``salt-key`` command makes use of this outputter to format its output.
 from __future__ import absolute_import
 
 # Import salt libs
-import salt.utils
 import salt.output
 from salt.utils.locales import sdecode
+import salt.utils.color
 
 
 def output(data, **kwargs):  # pylint: disable=unused-argument
@@ -18,7 +18,7 @@ def output(data, **kwargs):  # pylint: disable=unused-argument
     Read in the dict structure generated by the salt key API methods and
     print the structure.
     '''
-    color = salt.utils.get_colors(
+    color = salt.utils.color.get_colors(
             __opts__.get('color'),
             __opts__.get('color_theme'))
     strip_colors = __opts__.get('strip_colors', True)
