+++ salt/output/nested.py
@@ -29,9 +29,9 @@ from numbers import Number
 
 # Import salt libs
 import salt.output
+import salt.utils.color
 import salt.utils.locales
 import salt.utils.odict
-from salt.utils import get_colors
 from salt.ext.six import string_types
 
 
@@ -41,7 +41,7 @@ class NestDisplay(object):
     '''
     def __init__(self):
         self.__dict__.update(
-            get_colors(
+            salt.utils.color.get_colors(
                 __opts__.get('color'),
                 __opts__.get('color_theme')
             )
