+++ salt/utils/__init__.py
@@ -126,7 +126,6 @@ import salt.utils.dictupdate
 import salt.utils.versions
 import salt.version
 from salt.utils.decorators.jinja import jinja_filter
-from salt.textformat import TextFormat
 from salt.exceptions import (
     CommandExecutionError, SaltClientError,
     CommandNotFoundError, SaltSystemExit,
@@ -136,83 +135,6 @@ from salt.exceptions import (
 
 log = logging.getLogger(__name__)
 _empty = object()
-
-
-def get_color_theme(theme):
-    '''
-    Return the color theme to use
-    '''
-    # Keep the heavy lifting out of the module space
-    import yaml
-    if not os.path.isfile(theme):
-        log.warning('The named theme {0} if not available'.format(theme))
-
-    # Late import to avoid circular import.
-    import salt.utils.files
-    try:
-        with salt.utils.files.fopen(theme, 'rb') as fp_:
-            colors = yaml.safe_load(fp_.read())
-            ret = {}
-            for color in colors:
-                ret[color] = '\033[{0}m'.format(colors[color])
-            if not isinstance(colors, dict):
-                log.warning('The theme file {0} is not a dict'.format(theme))
-                return {}
-            return ret
-    except Exception:
-        log.warning('Failed to read the color theme {0}'.format(theme))
-        return {}
-
-
-def get_colors(use=True, theme=None):
-    '''
-    Return the colors as an easy to use dict.  Pass `False` to deactivate all
-    colors by setting them to empty strings.  Pass a string containing only the
-    name of a single color to be used in place of all colors.  Examples:
-
-    .. code-block:: python
-
-        colors = get_colors()  # enable all colors
-        no_colors = get_colors(False)  # disable all colors
-        red_colors = get_colors('RED')  # set all colors to red
-    '''
-
-    colors = {
-        'BLACK': TextFormat('black'),
-        'DARK_GRAY': TextFormat('bold', 'black'),
-        'RED': TextFormat('red'),
-        'LIGHT_RED': TextFormat('bold', 'red'),
-        'GREEN': TextFormat('green'),
-        'LIGHT_GREEN': TextFormat('bold', 'green'),
-        'YELLOW': TextFormat('yellow'),
-        'LIGHT_YELLOW': TextFormat('bold', 'yellow'),
-        'BLUE': TextFormat('blue'),
-        'LIGHT_BLUE': TextFormat('bold', 'blue'),
-        'MAGENTA': TextFormat('magenta'),
-        'LIGHT_MAGENTA': TextFormat('bold', 'magenta'),
-        'CYAN': TextFormat('cyan'),
-        'LIGHT_CYAN': TextFormat('bold', 'cyan'),
-        'LIGHT_GRAY': TextFormat('white'),
-        'WHITE': TextFormat('bold', 'white'),
-        'DEFAULT_COLOR': TextFormat('default'),
-        'ENDC': TextFormat('reset'),
-    }
-    if theme:
-        colors.update(get_color_theme(theme))
-
-    if not use:
-        for color in colors:
-            colors[color] = ''
-    if isinstance(use, six.string_types):
-        # Try to set all of the colors to the passed color
-        if use in colors:
-            for color in colors:
-                # except for color reset
-                if color == 'ENDC':
-                    continue
-                colors[color] = colors[use]
-
-    return colors
 
 
 def get_context(template, line, num_lines=5, marker=None):
@@ -3435,3 +3357,49 @@ def kwargs_warn_until(kwargs,
         stacklevel=stacklevel,
         _version_info_=_version_info_,
         _dont_call_warnings=_dont_call_warnings)
+
+
+def get_color_theme(theme):
+    '''
+    Return the color theme to use
+
+    .. deprecated:: Oxygen
+    '''
+    # Late import to avoid circular import.
+    import salt.utils.color
+    import salt.utils.versions
+
+    salt.utils.versions.warn_until(
+        'Neon',
+        'Use of \'salt.utils.get_color_theme\' detected. This function has '
+        'been moved to \'salt.utils.color.get_color_theme\' as of Salt '
+        'Oxygen. This warning will be removed in Salt Neon.'
+    )
+    return salt.utils.color.get_color_theme(theme)
+
+
+def get_colors(use=True, theme=None):
+    '''
+    Return the colors as an easy to use dict.  Pass `False` to deactivate all
+    colors by setting them to empty strings.  Pass a string containing only the
+    name of a single color to be used in place of all colors.  Examples:
+
+    .. code-block:: python
+
+        colors = get_colors()  # enable all colors
+        no_colors = get_colors(False)  # disable all colors
+        red_colors = get_colors('RED')  # set all colors to red
+
+    .. deprecated:: Oxygen
+    '''
+    # Late import to avoid circular import.
+    import salt.utils.color
+    import salt.utils.versions
+
+    salt.utils.versions.warn_until(
+        'Neon',
+        'Use of \'salt.utils.get_colors\' detected. This function has '
+        'been moved to \'salt.utils.color.get_colors\' as of Salt '
+        'Oxygen. This warning will be removed in Salt Neon.'
+    )
+    return salt.utils.color.get_colors(use=use, theme=theme)
