+++ tests/unit/utils/test_color.py
@@ -1,0 +1,27 @@
+# -*- coding: utf-8 -*-
+'''
+Unit tests for salt.utils.color.py
+'''
+
+# Import python libs
+from __future__ import absolute_import
+
+# Import Salt Testing libs
+from tests.support.unit import TestCase
+
+# Import Salt libs
+import salt.utils.color
+
+
+class ColorUtilsTestCase(TestCase):
+
+    def test_get_colors(self):
+        ret = salt.utils.color.get_colors()
+        self.assertEqual('\x1b[0;37m', str(ret['LIGHT_GRAY']))
+
+        ret = salt.utils.color.get_colors(use=False)
+        self.assertDictContainsSubset({'LIGHT_GRAY': ''}, ret)
+
+        ret = salt.utils.color.get_colors(use='LIGHT_GRAY')
+        # LIGHT_YELLOW now == LIGHT_GRAY
+        self.assertEqual(str(ret['LIGHT_YELLOW']), str(ret['LIGHT_GRAY']))
